Package and utilities renaming
