+++ app/src/main/java/com/fjoglar/lyricly/core/SingleLiveEvent.java
@@ -15,7 +15,7 @@
  */
 
 
-package com.fjoglar.lyricly.util;
+package com.fjoglar.lyricly.core;
 
 import android.util.Log;
