+++ app/src/main/java/com/fjoglar/lyricly/core/extensions/Views.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.fjoglar.lyricly.util.extensions
+package com.fjoglar.lyricly.core.extensions
 
 import android.view.View
 import android.view.animation.AccelerateDecelerateInterpolator
