+++ app/src/main/java/com/fjoglar/lyricly/core/navigation/Navigator.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018 Felipe Joglar Santos
+ * Copyright 2019 Felipe Joglar Santos
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com.fjoglar.lyricly.util.usecases;
+package com.fjoglar.lyricly.core.navigation
 
-import com.fjoglar.lyricly.data.SongsDataSource;
+import android.content.Context
+import android.content.Intent
 
-import io.reactivex.Flowable;
+object Navigator {
 
-public interface FlowableUseCase<P, R> {
-
-    Flowable<R> execute(SongsDataSource songsDataSource, P parameter);
+    fun showNotificationPermissionSettings(context: Context) = context.startActivity(
+        Intent("android.settings.ACTION_NOTIFICATION_LISTENER_SETTINGS")
+    )
 }
