+++ app/src/main/java/com/fjoglar/lyricly/core/schedulers/SchedulerProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.fjoglar.lyricly.util.schedulers;
+package com.fjoglar.lyricly.core.schedulers;
 
 import androidx.annotation.Nullable;
