+++ app/src/main/java/com/fjoglar/lyricly/core/utils/AnimUtil.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2019 Felipe Joglar Santos
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.fjoglar.lyricly.core.utils;
+
+
+import android.content.Context;
+import android.view.animation.AnimationUtils;
+import android.view.animation.Interpolator;
+
+/**
+ * Utility methods for working with animations.
+ */
+public class AnimUtil {
+
+    private AnimUtil() {
+    }
+
+    private static Interpolator fastOutSlowIn;
+
+    public static Interpolator getFastOutSlowInInterpolator(Context context) {
+        if (fastOutSlowIn == null) {
+            fastOutSlowIn = AnimationUtils.loadInterpolator(context,
+                    android.R.interpolator.fast_out_slow_in);
+        }
+        return fastOutSlowIn;
+    }
+}
