+++ app/src/main/java/com/fjoglar/lyricly/currentlyplayingsong/CurrentlyPlayingSongNotification.kt
@@ -29,19 +29,19 @@ import androidx.core.app.NotificationManagerCompat
 import com.fjoglar.lyricly.R
 import com.fjoglar.lyricly.data.model.Song
 import com.fjoglar.lyricly.song.SongActivity
-import com.fjoglar.lyricly.util.ImageLoaderHelper
+import com.fjoglar.lyricly.core.utils.ImageLoader
 
 const val CHANNEL_ID = "Lyricly currently playing song channel"
 const val NOTIFICATION_ID = 8888
 
 fun launchCurrentlyPlayingSongNotification(applicationContext: Context, song: Song) {
     createNotificationChannel(applicationContext)
-    ImageLoaderHelper.loadBitmapFromNetwork(
+    ImageLoader.loadBitmapFromNetwork(
         applicationContext,
         applicationContext.getString(
             R.string.album_image_url,
             song.albumId,
-            ImageLoaderHelper.IMAGE_SIZE_BIG
+            ImageLoader.IMAGE_SIZE_BIG
         ),
         16
     ) {
