+++ app/src/main/java/com/fjoglar/lyricly/song/DeleteSongFromFavoriteUseCase.java
@@ -18,7 +18,7 @@ package com.fjoglar.lyricly.song;
 
 import com.fjoglar.lyricly.data.SongsDataSource;
 import com.fjoglar.lyricly.data.model.Song;
-import com.fjoglar.lyricly.util.usecases.CompletableUseCase;
+import com.fjoglar.lyricly.core.usecases.CompletableUseCase;
 
 import io.reactivex.Completable;
