+++ app/src/main/java/com/fjoglar/lyricly/song/SongFragment.java
@@ -32,8 +32,8 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.fjoglar.lyricly.R;
 import com.fjoglar.lyricly.data.model.Song;
-import com.fjoglar.lyricly.util.ImageLoaderHelper;
-import com.fjoglar.lyricly.util.Injection;
+import com.fjoglar.lyricly.core.utils.ImageLoader;
+import com.fjoglar.lyricly.core.Injection;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
 
@@ -185,10 +185,10 @@ public class SongFragment extends Fragment {
         mSong = song;
 
         String albumId = mSong.getAlbumId();
-        String imageUrl = getString(R.string.album_image_url, albumId, ImageLoaderHelper.IMAGE_SIZE_BIG);
-        String thumbUrl = getString(R.string.album_image_url, albumId, ImageLoaderHelper.IMAGE_SIZE_SMALL);
+        String imageUrl = getString(R.string.album_image_url, albumId, ImageLoader.IMAGE_SIZE_BIG);
+        String thumbUrl = getString(R.string.album_image_url, albumId, ImageLoader.IMAGE_SIZE_SMALL);
 
-        ImageLoaderHelper.loadWithThumb(mImageViewSongCover, imageUrl, thumbUrl);
+        ImageLoader.loadWithThumb(mImageViewSongCover, imageUrl, thumbUrl);
         mTextViewSongTitle.setText(mSong.getName());
         mTextViewSongArtist.setText(mSong.getArtistName());
         mTextViewSongLyrics.setText(mSong.getLyrics());
