+++ app/src/main/java/com/fjoglar/lyricly/songs/recent/RecentSongsViewModel.java
@@ -23,8 +23,8 @@ import androidx.lifecycle.ViewModel;
 import com.fjoglar.lyricly.data.SongsDataSource;
 import com.fjoglar.lyricly.songs.SongsResponse;
 import com.fjoglar.lyricly.songs.SongsViewModel;
-import com.fjoglar.lyricly.util.Injection;
-import com.fjoglar.lyricly.util.schedulers.SchedulerProvider;
+import com.fjoglar.lyricly.core.Injection;
+import com.fjoglar.lyricly.core.schedulers.SchedulerProvider;
 
 import io.reactivex.disposables.CompositeDisposable;
